Hotels: Prepare payment screen for stay22 integration
